Add files via upload
Our project is about employee information system that mainly include certain functionality for the user(mainly for the manager).
allows the user(manager)   -To Register to the system and 
                                           -To login to the system 
The register functionality is used for the user that access the system as a manager. when the manager askes to register. The system will prompt the user to insert a username and a password which will be written to a file to save the information inserted. after registering to the system the user can login to the system that will lead the the manager to access the employee information such as - To add employee information
                                                                       -To update employee information 
                                                                       -To delete employee information
                                                                       -To display all employee information 
                                                                       -To search for specific employee information